Separate date and datetime components
