+++ packages/twenty-front/src/modules/object-record/record-field/form-types/components/FormDateFieldInput.tsx
@@ -1,0 +1,26 @@
+import { FormDateTimeFieldInput } from '@/object-record/record-field/form-types/components/FormDateTimeFieldInput';
+import { VariablePickerComponent } from '@/object-record/record-field/form-types/types/VariablePickerComponent';
+
+type FormDateFieldInputProps = {
+  label?: string;
+  defaultValue: string | undefined;
+  onPersist: (value: string | null) => void;
+  VariablePicker?: VariablePickerComponent;
+};
+
+export const FormDateFieldInput = ({
+  label,
+  defaultValue,
+  onPersist,
+  VariablePicker,
+}: FormDateFieldInputProps) => {
+  return (
+    <FormDateTimeFieldInput
+      dateOnly
+      label={label}
+      defaultValue={defaultValue}
+      onPersist={onPersist}
+      VariablePicker={VariablePicker}
+    />
+  );
+};
